initial impl of inertialization

+++ Source/InceptionEngine/RunTime/Animation/Animation.cpp
@@ -41,16 +41,16 @@ namespace inceptionengine
 
 		time = std::fmodf(time, mDuration);
 
-		int count = static_cast<int>(time * 30.0f);
+		int count = static_cast<int>(static_cast<double>(time) * 30.0);
 
 		if(count >= mBoneTransforms.size() - 1)
 			return AnimPose(mBoneTransforms.back(), mBoneLclTranslVelocities.back(), mBoneGlobalTranslVelocities.back(),
 				mBoneLclAngularVelocities.back(), mBoneGlobalAngularVelocities.back());
 
-		float countTime1 = static_cast<float>(count) / 30.0f;
-		float countTime2 = countTime1 + (1.0f / 30.0f);
+		double countTime1 = static_cast<double>(count) / 30.0;
+		double countTime2 = countTime1 + (1.0 / 30.0);
 		assert(time >= countTime1 && time <= countTime2);
-		float alpha = 1.0f - ((time - countTime1) / (1.0f / 30.0f)); //weights of the first pose
+		float alpha = static_cast<float>(1.0 - ((static_cast<double>(time) - countTime1) / (1.0 / 30.0))); //weights of the first pose
 
 		return AnimPose(LinearInterpolate(mBoneTransforms[count], mBoneTransforms[count + 1], alpha),
 			mBoneLclTranslVelocities[count], mBoneGlobalTranslVelocities[count],
